+++ R/plot.R
@@ -27,20 +27,35 @@ make_label = function(x, i = 1) {
 #' @param reset logical; if \code{FALSE}, keep the plot in a mode that allows adding further map elements; if \code{TRUE} restore original mode after plotting
 #' @param box_col color for box around sub-plots; use \code{0} to suppress plotting of boxes around sub-plots.
 #' @param center_time logical; if \code{TRUE}, sub-plot titles will show the center of time intervals, otherwise their start
-#' @param hook NULL or function; hook function that will be called on every sub-plot.
+#' @param hook NULL or function; hook function that will be called on every sub-plot; see examples.
 #' @param mfrow length-2 integer vector with nrows, ncolumns of a composite plot, to override the default layout
 #' @details
-#' Downsampling: a value for \code{downsample} of 0: no downsampling, 1: after every dimension value (pixel/line/band), one value is skipped (half of the original resolution), 2: after every dimension value, 2 values are skipped (one third of the original resolution), etc.
+#' Downsampling: a value for \code{downsample} of 0: no downsampling, 1: after every dimension value (pixel/line/band), one value is skipped (half of the original resolution), 2: after every dimension value, 2 values are skipped (one third of the original resolution), etc. If \code{downsample} is \code{TRUE} or a length 1 numeric vector, downsampling is only applied to the raster [x] and [y] dimensions.
 #'
 #' To remove unused classes in a categorical raster, use the \link[base]{droplevels} function.
 #'
 #' When bitmaps show visual artefacts (Moiré effects), make sure that device \link{png} is used rather than \code{ragg::agg_png} as the latter uses antialiasing for filled polygons which causes this; see also https://github.com/r-spatial/stars/issues/573 .
 #' @export
+#' @examples
+#' st_bbox(L7_ETMs) |> st_as_sfc() |> st_centroid() |> st_coordinates() -> pt
+#' hook1 = function() {
+#'     text(pt[,"X"], pt[,"Y"], "foo", col = 'orange', cex = 2)
+#' }
+#' plot(L7_ETMs, hook = hook1)
+#' x = st_set_dimensions(L7_ETMs, 3, paste0("B_", 1:6))
+#' hook2 = function(..., row, col, nr, nrow, ncol, value, bbox) {
+#'    str = paste0("row ", row, "/", nrow, ", col ", col, "/", ncol, "\nnr: ", nr, " value: ", value)
+#'    bbox |> st_as_sfc() |> st_centroid() |> st_coordinates() -> pt
+#'    text(pt[,"X"], pt[,"Y"], str, col = 'red', cex = 2)
+#' }
+#' plot(x, hook = hook2, col = grey(c(.2,.25,.3,.35)))
 plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes = FALSE,
 		downsample = TRUE, nbreaks = 11, breaks = "quantile", col = grey(1:(nbreaks-1)/nbreaks),
 		key.pos = get_key_pos(x, ...), key.width = lcm(1.8), key.length = 0.618,
 		reset = TRUE, box_col = grey(.8), center_time = FALSE, hook = NULL, mfrow = NULL) {
 
+	if (!missing(y))
+		stop("y argument should be missing")
 	flatten = function(x, i) { # collapse all non-x/y dims into one, and select "layer" i
 		d = st_dimensions(x)
 		dxy = attr(d, "raster")$dimensions
@@ -76,6 +91,27 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 	if (is.factor(x[[1]]) && any(is.na(levels(x[[1]]))))
 		x = droplevels(x) # https://github.com/r-spatial/stars/issues/339
 
+	zlim = if (join_zlim && is.null(dots$rgb))
+			range(unclass(x[[1]]), na.rm = TRUE)
+		else
+			rep(NA_real_, 2)
+
+	# downsample if raster:
+	if (has_raster(x) && !identical(downsample, FALSE)) {
+		dxy = attr(st_dimensions(x), "raster")$dimensions
+		loop = setdiff(names(dim(x)), dxy) # dimension (name) over which we loop, if any
+		x = aperm(x, c(dxy, loop))
+		dims = dim(x)
+		if (isTRUE(downsample)) {
+			n = dims * 0 # keep names
+			n[dxy] = get_downsample(dims, rgb = is.numeric(dots$rgb))
+			x = st_downsample(x, n)
+		} else if (is.numeric(downsample)) {
+			x = st_downsample(x, downsample)
+		} 
+	}
+
+	# find breaks:
 	if (join_zlim && !is.character(x[[1]]) && is.null(dots$rgb)) {
 		breaks = if (is.factor(x[[1]]))
 					seq(.5, length.out = length(levels(x[[1]])) + 1)
@@ -93,27 +129,8 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 	if (isTRUE(dots$logz) && !((has_raster(x) && (is_curvilinear(x) || has_rotate_or_shear(x))) || has_sfc(x)))
 		x = log10(x) # otherwise, defer log-transforming to sf::plot.sf
 
-	if (!missing(y))
-		stop("y argument should be missing")
 	if (has_raster(x)) {
-		dxy = attr(st_dimensions(x), "raster")$dimensions
-		loop = setdiff(names(dim(x)), dxy) # dimension (name) over which we loop, if any
-		x = aperm(x, c(dxy, loop))
-		zlim = if (join_zlim && is.null(dots$rgb))
-				range(unclass(x[[1]]), na.rm = TRUE)
-			else
-				rep(NA_real_, 2)
 		dims = dim(x)
-		x = if (isTRUE(downsample)) {
-				n = dims * 0 # keep names
-				n[dxy] = get_downsample(dims, rgb = is.numeric(dots$rgb))
-				st_downsample(x, n)
-			} else if (is.numeric(downsample)) {
-				st_downsample(x, downsample)
-			} else
-				x
-		dims = dim(x) # may have changed by st_downsample
-
 		if (length(dims) == 2 || dims[3] == 1 || (!is.null(dots$rgb) && is.numeric(dots$rgb))) { ## ONE IMAGE:
 			# set up key region
 			values = structure(x[[1]], dim = NULL) # array -> vector
@@ -175,8 +192,19 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 					else # user-defined
 						title(paste(main, format(labels[i])))
 				}
-				if (!is.null(hook) && is.function(hook))
-					hook()
+				if (!is.null(hook) && is.function(hook)) {
+					if (is.null(formals(hook)))
+						hook()
+					else {
+						nc = lt$mfrow[[2]] # nr of columns in the plot
+						hook(row = ((i - 1) %/% nc) + 1,
+							 col = ((i - 1)  %% nc) + 1, 
+							 nrow = lt$mfrow[[1]],
+							 ncol = nc,
+							 nr = i, value = labels[i],
+							 bbox = st_bbox(x))
+					}
+				}
 				box(col = box_col)
 			}
 			for (i in seq_len(prod(lt$mfrow) - dims[3])) # empty panels:
@@ -252,7 +280,7 @@ get_breaks = function(x, breaks, nbreaks, logz = NULL) {
 #' tif = system.file("tif/L7_ETMs.tif", package = "stars")
 #' x = read_stars(tif)
 #' image(x, col = grey((3:9)/10))
-#' image(x, rgb = c(1,3,5)) # rgb composite
+#' image(x, rgb = c(1,3,5)) # false color composite
 image.stars = function(x, ..., band = 1, attr = 1, asp = NULL, rgb = NULL,
 		maxColorValue = ifelse(inherits(rgb, "data.frame"), 255, max(x[[attr]], na.rm = TRUE)),
 		xlab = if (!axes) "" else names(d)[1], ylab = if (!axes) "" else names(d)[2],
@@ -503,6 +531,12 @@ st_rgb <- function (x,
 					maxColorValue = 255L,
 					probs = c(0, 1),
 					stretch = NULL) {
+
+	if (inherits(x, "stars_proxy"))
+		return(collect(x, match.call(), "st_rgb", 
+					   c("x", "dimension", "use_alpha", "maxColorValue", "probs", "stretch"), env = environment())) # RETURNS!!
+
+	# if not stars_proxy:
 	if (is.character(dimension))
 		dimension = match(dimension, names(dim(x)))
 	stopifnot(is.numeric(dimension), length(dimension) == 1)