+++ R/read.R
@@ -25,6 +25,10 @@ is_functions = function(x) {
 	is.function(x) || all(sapply(x, is.function))
 }
 
+is_big = function(x, ..., sub = sub, n_proxy) {
+	prod(dim(read_stars(x, ..., sub = sub, proxy = TRUE, quiet = TRUE))) > n_proxy
+}
+
 # strip common start & end parts of x, if remainder != ""
 unique_part = function(x, prefix) {
 	stopifnot(is.character(x), length(x) > 1)
@@ -55,8 +59,7 @@ unique_part = function(x, prefix) {
 #' @param RasterIO list with named parameters for GDAL's RasterIO, to further control the extent, resolution and bands to be read from the data source; see details.
 #' @param proxy logical; if \code{TRUE}, an object of class \code{stars_proxy} is read which contains array
 #' metadata only; if \code{FALSE} the full array data is read in memory. Always \code{FALSE} for curvilinear girds.
-#' If not set, defaults to \code{TRUE} when the number of cells to be read is larger than \code{options(stars.n_proxy},
-#' or to 1e8 if that option was not set.
+#' If set to a number, defaults to \code{TRUE} when the number of cells to be read is larger than that number.
 #' @param curvilinear length two character vector with names of subdatasets holding longitude and latitude values for all raster cells, or named length 2 list holding longitude and latitude matrices; the names of this list should correspond to raster dimensions referred to
 #' @param normalize_path logical; if \code{FALSE}, suppress a call to \link{normalizePath} on \code{.x}
 #' @param RAT character; raster attribute table column name to use as factor levels
@@ -112,10 +115,9 @@ unique_part = function(x, prefix) {
 #'   plot(st_as_sfc(red, as_points = FALSE), add = TRUE, border = 'green', lwd = 2)
 #' }
 #' file.remove(tmp)
-read_stars = function(.x, ..., options = character(0), driver = character(0),
-		sub = TRUE, quiet = FALSE, NA_value = NA_real_, along = NA_integer_,
-		RasterIO = list(), proxy = is_functions(.x) || (!length(curvilinear) &&
-				is_big(.x, sub = sub, driver = driver, normalize_path = normalize_path, ...)),
+read_stars = function(.x, sub = TRUE, ..., options = character(0), 
+		driver = character(0), quiet = FALSE, NA_value = NA_real_, along = NA_integer_,
+		RasterIO = list(), proxy = getOption("stars.n_proxy") %||% 1.e8,
 		curvilinear = character(0), normalize_path = TRUE, RAT = character(0),
 		tolerance = 1e-10, exclude = "", shorten = TRUE) {
 
@@ -123,6 +125,12 @@ read_stars = function(.x, ..., options = character(0), driver = character(0),
 	if (!is.list(.x) && length(.x) > 1 && is.null(names(.x)) && !is_false(shorten) &&
 				any((short_names <- unique_part(.x, shorten)) != .x))
 		names(.x) = short_names
+	if (is.numeric(proxy)) {
+		if (is_functions(.x))
+			proxy = TRUE
+		else if (length(curvilinear))
+			proxy = FALSE
+	}
 
 	x = if (is.list(.x)) {
 			f = function(y, np) enc2utf8char(maybe_normalizePath(y, np))
@@ -160,16 +168,16 @@ read_stars = function(.x, ..., options = character(0), driver = character(0),
 	if (length(curvilinear) && proxy)
 		warning("proxy = TRUE may not work for curvilinear rasters")
 
-	if (length(x) > 1) { # loop over data sources and RETURNS:
+	if (length(x) > 1) { # loop over data sources, read, c() all, and RETURN:
 		ret = lapply(x, read_stars, options = options, driver = driver, sub = sub, quiet = quiet,
 			NA_value = NA_value, RasterIO = as.list(RasterIO), proxy = proxy, curvilinear = curvilinear,
 			along = if (length(along) > 1) along[-1] else NA_integer_, normalize_path = normalize_path)
 		return(do.call(c, append(ret, list(along = along, tolerance = tolerance))))
 	}
 
-	# else:
+	# else: do it --
 	data = sf::gdal_read(get_names(x),
-		options = options, driver = driver, read_data = !proxy,
+		options = options, driver = driver, read_data = ifelse(is.logical(proxy), !proxy, proxy),
 		NA_value = NA_value, RasterIO_parameters = as.list(RasterIO))
 	if (!is.null(data$default_geotransform) && data$default_geotransform == 1) {
 		## we have the 0 1 0 0 0 1 transform indicated
@@ -227,6 +235,8 @@ read_stars = function(.x, ..., options = character(0), driver = character(0),
 				warning("only one array present: argument 'sub' will be ignored")
 		}
 		meta_data = structure(data, data = NULL) # take meta_data only
+		if (is.numeric(proxy) && !is.null(data$proxy))
+			proxy = data$proxy
 		data = if (proxy)
 				.x
 			else
@@ -273,9 +283,11 @@ read_stars = function(.x, ..., options = character(0), driver = character(0),
 				# select attribute table:
 				which.at = which(lengths(at) > 0)[1]
 				which.column = if (length(RAT))
-						RAT
+						RAT[1]
 					else
 						which(sapply(at[[which.at]], class) == "character")[1]
+				if (length(which.column) == 0)
+					which.column = 1
 				labels = at = at[[ which.at ]][[ which.column ]]
 				levels = 0:(length(at) - 1)
 				if (length(exclude)) {
@@ -287,8 +299,10 @@ read_stars = function(.x, ..., options = character(0), driver = character(0),
 				} else
 					ex = rep(FALSE, length(levels))
 			} else {
-				levels = sort(unique(data))
+				levels = sort(unique(as.vector(data)))
 				labels = as.character(levels)
+				if (!is.null(co))
+					co = co[levels + 1]
 				ex = rep(FALSE, length(levels))
 			}
 			# f = factor(as.vector(data), levels = levels, labels = labels) # is too costly;
@@ -321,9 +335,10 @@ read_stars = function(.x, ..., options = character(0), driver = character(0),
 				x
 		name_x = tail(strsplit(name_x, '[\\\\/:]+')[[1]], 1)
 		ret = if (proxy) { # no data present, subclass of "stars":
-				st_stars_proxy(setNames(list(x), names(.x) %||% name_x),
-					create_dimensions_from_gdal_meta(dims, meta_data), NA_value = NA_value,
-					resolutions = NULL, RasterIO = RasterIO, file_dim = dims)
+				dims = create_dimensions_from_gdal_meta(dims, meta_data)
+				st_stars_proxy(setNames(list(x), names(.x) %||% name_x), dims,
+					NA_value = NA_value, resolutions = NULL, RasterIO = RasterIO, 
+					file_dim = matrix(dim(dims), 1, dimnames = list(NULL, names(dims))))
 			} else
 				st_stars(setNames(list(data), names(.x) %||% name_x),
 					create_dimensions_from_gdal_meta(dim(data), meta_data))
@@ -333,15 +348,6 @@ read_stars = function(.x, ..., options = character(0), driver = character(0),
 		else
 			ret
 	}
-}
-
-#' @export
-#' @name read_stars
-#' @param x object to be read with \link{read_stars}
-#' @param n_proxy integer; number of cells above which .x will be read as stars
-#' proxy object, i.e. not as in-memory arrays but left on disk
-is_big = function(x, ..., sub = sub, n_proxy = options("stars.n_proxy")[[1]] %||% 1.e8) {
-	prod(dim(read_stars(x, ..., sub = sub, proxy = TRUE, quiet = TRUE))) > n_proxy
 }
 
 get_data_units = function(data) {