+++ R/sf.R
@@ -48,6 +48,8 @@ st_xy2sfc = function(x, as_points, ..., na.rm = TRUE) {
 	dxy = attr(d, "raster")$dimensions
 	xy_pos = match(dxy, names(d))
 	stopifnot(all(xy_pos == 1:2))
+	if (missing(as_points) && isTRUE(d[[dxy[1]]]$point) && isTRUE(d[[dxy[2]]]$point))
+		as_points = TRUE
 
 	# find which records are NA for all attributes:
 	a = abind(x, along = length(dim(x)) + 1)
@@ -165,7 +167,9 @@ st_as_sf.stars = function(x, ..., as_points = FALSE, merge = FALSE, na.rm = TRUE
 				warning("working on the first sfc dimension only") # FIXME: this probably only works for 2D arrays, now
 			other_dim = setdiff(seq_along(dim(x)), ix[1])
 			sfc = st_dimensions(x)[[ ix[1] ]]$values
-			other_values = st_dimensions(x)[[ other_dim[1] ]]$values
+			# other_values = st_dimensions(x)[[ other_dim[1] ]]$values
+			other_values = lapply(st_dimensions(x)[other_dim], function(x) x$values)
+			varnames = apply(do.call(expand.grid, other_values), 1, paste, collapse = ".")
 			un_dim = function(x) { # remove a dim attribute from data.frame columns
 				for (i in seq_along(x))
 					x[[i]] = structure(x[[i]], dim = NULL)
@@ -177,8 +181,8 @@ st_as_sf.stars = function(x, ..., as_points = FALSE, merge = FALSE, na.rm = TRUE
 	
 			if (length(dim(x)) == 1) # one-dimensional cube...
 				names(df) = names(x)
-			else if (!is.null(other_values) && length(other_values) == ncol(df))
-				names(df) = other_values
+			else if (length(varnames) == ncol(df))
+				names(df) = varnames
 			else if (length(unique(names(df))) < ncol(df) && length(names(dfs)) == ncol(df)) # I hate this
 				names(df) = names(dfs)
 			else { # another exception... time as second dimension